organized extra function not used
