+++ client/src/__test__/convertNumberHrsToTime.test.js
@@ -1,8 +1,8 @@
-import { convertNumberHrsToTime } from "../helpers/convertNumberHrsToTime";
+import { convertNumberHrsToTime } from "../helpers/notused_yet/convertNumberHrsToTime";
 
 test("Convert the hours work to a workable time input", () => {
   expect(convertNumberHrsToTime("8")).toStrictEqual(['08','0']);
   expect(convertNumberHrsToTime("8.5")).toStrictEqual(['08','5']);
   expect(convertNumberHrsToTime("10")).toStrictEqual(['10','0']);
-  expect(convertNumberHrsToTime("10.1")).toStrictEqual(['10','1']);
+  expect(convertNumberHrsToTime("10.1")).toStrictEqual(['10', '1']);
 });
