--- client/src/helpers/companyRecommendedTime.js
@@ -1,7 +1,0 @@
-const companyRecommendedTime = {
-  punchIn: "08:00",
-  punchOut: "16:30",
-  lunchTime: "12:00",
-  hrToEat: 0.5,
-};
-export { companyRecommendedTime };
