--- client/src/helpers/flexTimeScale.js
@@ -1,60 +1,0 @@
-const flexScale = [
-  { id: 1, mins: ["00", "01",'02'], round: "00", percentHour: 0.0 },
-  {
-    id: 2,
-    mins: ["03", "04", "05", "06", "07", "09"],
-    round: "06",
-    percentHour: 0.1,
-  },
-  {
-    id: 3,
-    mins: ["09", "10", "11", "12", "13", "14"],
-    round: "12",
-    percentHour: 0.2,
-  },
-  {
-    id: 4,
-    mins: ["15", "16", "17", "18", "19", "20"],
-    round: "18",
-    percentHour: 0.3,
-  },
-  {
-    id: 5,
-    mins: ["21", "22", "23", "24", "25", "26"],
-    round: "24",
-    percentHour: 0.4,
-  },
-  {
-    id: 6,
-    mins: ["27", "28", "29", "30", "31", "32"],
-    round: "30",
-    percentHour: 0.5,
-  },
-  {
-    id: 7,
-    mins: ["33", "34", "35", "36", "37", "38"],
-    round: "36",
-    percentHour: 0.6,
-  },
-  {
-    id: 8,
-    mins: ["39", "40", "41", "42", "43", "44"],
-    round: "42",
-    percentHour: 0.7,
-  },
-  {
-    id: 9,
-    mins: ["45", "46", "47", "48", "49", "50"],
-    round: "48",
-    percentHour: 0.8,
-  },
-  {
-    id: 10,
-    mins: ["51", "52", "53", "54", "55", "56"],
-    round: "54",
-    percentHour: 0.9,
-  },
-  { id: 11, mins: ["57", "58", "59"], round: "00", percentHour: 0.0 },
-];
-
-export { flexScale };
