+++ client/src/helpers/notused_yet/flexMinutePercentHr.js
@@ -1,10 +1,11 @@
-import { flexScale } from "./flexTimeScale";
+import { flexScale } from "../standardDefaults/flexTimeScale";
 /**
  * Takes a minute from the user input the would send back a 
  * the percent of hour that the min clock in time falls into 
  * 
  * for punching out value it should have the first minute. 
  * for punching in it doesn't matter , cause can control it
+ * 
  * 
  * @param {sting} logMinute 
  */
