--- client/src/helpers/standardSettingForDay.js
@@ -1,11 +1,0 @@
-const standardSettingForDay = [
-  { id: 1, day: "Monday", isActive: true },
-  { id: 2, day: "Tuesday", isActive: true },
-  { id: 3, day: "Wednesday", isActive: true },
-  { id: 4, day: "Thursday", isActive: true },
-  { id: 5, day: "Friday", isActive: true },
-  { id: 6, day: "Saturday", isActive: false },
-  { id: 7, day: "Sunday", isActive: false },
-];
-
-export { standardSettingForDay };
